reorganize components directory
