+++ client/src/components/ColorWall/NoPostsComment.js
