+++ client/src/components/HomePage/ColorGrid/ColorGrid.js
