+++ client/src/components/HomePage/ColorGrid/style.css
