+++ client/src/components/index.js
@@ -1,6 +1,28 @@
+// HOME PAGE:
+export { default as ColorButtons } from "./HomePage/ColorButtons/ColorButtons";
+export { default as ColorGrid } from "./HomePage/ColorGrid/ColorGrid";
+export { default as Welcome } from "./HomePage/Welcome/Welcome";
+export { default as LoginButton } from "./HomePage/LoginButton";
+export { default as SignupButton } from "./HomePage/SignupButton";
+
+// COLORWALL COMPONENTS (ButtonRow, Modals, etc.):
+export { default as ButtonRow } from "./ColorWall/ButtonRow";
+export { default as ColorWallModal } from "./ColorWall/ColorWallModal";
+export { default as NoPostsComment } from "./ColorWall/NoPostsComment";
+
+// LOGIN/SIGNUP PAGES:
+export { default as LoginForm } from "./AuthForms/LoginForm";
+export { default as SignupForm } from "./AuthForms/SignupForm";
+
+// NAVIGATION BAR:
 export { default as Navigation } from "./Navigation/Navigation";
-export { default as PostTable } from "./PostTable";
-export { default as PostForm } from "./PostForm";
-export { default as LoginForm } from "./LoginForm";
-export { default as SignupForm } from "./SignupForm";
-export { default as Error } from "./Error";
+export { default as TempDrawer } from "./Navigation/TempDrawer";
+
+// FORUM PAGE:
+export { default as MessageCard } from "./ForumPage/MessageCard";
+export { default as PostForm } from "./ForumPage/PostForm";
+export { default as PostTable } from "./ForumPage/PostTable";
+
+// ERROR:
+export { default as Error } from "./Error/Error";
+export { default as PhotoUrl } from "./Error/PhotoUrl";
